+++ webviewsrc/miopreview.ts
@@ -2,7 +2,7 @@ import { getState, setState, arrayToMap, subscribeNavigators, scrollToState, try
 import { DivDropdown } from "./util/dropdown";
 import { minBy } from "lodash";
 import { renderGridBoxCommon, GridBoxItem, GridBoxConnection } from "../src/util/hoi4gui/gridboxcommon";
-import { StyleTable, normalizeForStyle } from "../src/util/styletable";
+import { StyleTable } from "../src/util/styletable";
 import { applyCondition, ConditionItem } from "../src/hoiformat/condition";
 import { NumberPosition } from "../src/util/common";
 import { GridBoxType } from "../src/hoiformat/gui";